EHN: Add UAV example to docs
